course-selector/backend/requirements.txt
anthropic
fastapi
uvicorn
python-dotenv
pydantic